refactor: move prose-report-generator to review stage
The review stage (pipeline-executor) was importing generateProseReport
from src/stages/report/generators/, inverting the expected data-flow
direction where review produces data and report writes outputs.

Move prose-report-generator.ts into src/stages/review/ so the import
is within the same stage. The report stage's file-detection short-circuit
(existsSync on prose-report.md) is unchanged.
